third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/require-sri-for/require-sri-for-style-preload-blocked-expected.txt

CONSOLE ERROR: line 22: Refused to load the stylesheet 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/style-set-red.css' because 'require-sri-for' directive requires integrity attribute be present for all stylesheets.
This is a testharness.js-based test.
PASS Script that requires integrity executes and does not generate a violation report. 
PASS Makes sure that require-sri-for applies to preloaded resources. 
PASS Stylesheets without integrity do not load. 
Harness: the test ran to completion.
